component id u16

+++ mls-rs/src/group/application_data.rs
@@ -69,4 +69,4 @@ impl fmt::Debug for ComponentData {
     }
 }
 
-pub type ComponentId = u32;
+pub type ComponentId = u16;
