+++ mls-rs/src/group/commit.rs
@@ -923,7 +923,7 @@ where
 
         #[cfg(feature = "application_data")]
         let application_data = {
-            let mut data: BTreeMap<u32, Vec<Vec<u8>>> = BTreeMap::new();
+            let mut data: BTreeMap<u16, Vec<Vec<u8>>> = BTreeMap::new();
             for proposal in provisional_state
                 .applied_proposals
                 .app_ephemeral_proposals()
